requirements.txt
discord.py
asyncio
git+https://github.com/MysterialPy/TwitchIO.git